13/7 (again)
The RETURN button from the settings screen bothered me LOTS. First, the node I went to after that froze (I realized now that the whole thing was in 'pause mode'), and after that particular bug was fixed, I can't get over the fact that it'll only go back to MainMenu.tscn, as it was coded in the button to do so.

So I changed it. That SceneManager will now safe any scene you were in so that it can smoothly return to the scene before that, and I won't have to make it like I did the pause screen.