5/8
FIXED THE SWAP THING!!! It works now :D

Added a mechanic where the player has to put specific objects on the tables and/or killing all enemies before they can enter the next level.
Added a 'thank you' panel after the dialogue (if the player chooses to help the game world and thus, start the game).