18/8
OK SO. THIS IS A TRAIN WRECK, BUT I'VE FOUND THE BUGS!

But before that; let's hear what's working now:
- THE AMANDA COMPUTER CAN TALK!!! It's all fixed, the dialogue's been implanted, it's perfect.

Right. Back to the ~Bad News~
- While I was talking to Amanda the Computer, I tried pausing the screen. It worked-technically. But now the game thinks that I'm still interacting with the aforementioned computer... So I attempt to change it, of course. It worked-and then it doesn't. One thing led to another and the scenes ended up stacking above one another, the music, and everything just seemed to break apart. (There are some videos with me screaming internally in the background)

It kind of just spiraled out of control after that. I'm in the process of fixing it, but I'll probably just make the settings screen the way the pause screen is set up (with an addition of a colored-background; so it doesn't look like multitudes of scenes stacking on top of each other. It technically would be, but I'll just be fooling the peeps with a prettier view).

This is why we do testing, guys. To find these sneaky little bugs.