20/8
the settings works better now. I'll still have to add a bg so it doesn't clash with the scene below it; and the ordering of the pause screen one (the pause screen is still on top of it, why?).

changed up the second level's map (it was too big) and added at least a little structure to it.